Replace some occurences of int with NSInteger and unsigned with NSUInteger. This allows to recompile gui on 64 bit systems after the change to NSNotFound in base.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@34312 72102866-910b-0410-8b05-ffd578937521

+++ ColorPickers/GSNamedColorPicker.m
@@ -110,7 +110,7 @@
   NSColor *c = [color colorUsingColorSpaceName: NSNamedColorSpace];
   NSString *list;
   NSString *name;
-  unsigned int index;
+  NSUInteger index;
 
   if (c == nil)
     return;
