+++ Headers/AppKit/NSSlider.h
@@ -71,14 +71,14 @@
 // ticks
 - (BOOL) allowsTickMarkValuesOnly;
 - (double) closestTickMarkValueToValue: (double)aValue;
-- (int) indexOfTickMarkAtPoint: (NSPoint)point;
-- (int) numberOfTickMarks;
-- (NSRect) rectOfTickMarkAtIndex: (int)index;
+- (NSInteger) indexOfTickMarkAtPoint: (NSPoint)point;
+- (NSInteger) numberOfTickMarks;
+- (NSRect) rectOfTickMarkAtIndex: (NSInteger)index;
 - (void) setAllowsTickMarkValuesOnly: (BOOL)flag;
-- (void) setNumberOfTickMarks: (int)numberOfTickMarks;
+- (void) setNumberOfTickMarks: (NSInteger)numberOfTickMarks;
 - (void) setTickMarkPosition: (NSTickMarkPosition)position;
 - (NSTickMarkPosition) tickMarkPosition;
-- (double) tickMarkValueAtIndex: (int)index;
+- (double) tickMarkValueAtIndex: (NSInteger)index;
 #endif
 
 @end
